found error index on different image

+++ src/gpu/kernel.cuh
@@ -4,7 +4,7 @@
 
 struct Rgb {
     __host__ __device__ Rgb() {}
-    __host__ __device__ Rgb(float x, float y, float z) : r(x), g(y), b(z) {}
+    __host__ __device__ Rgb(double x, double y, double z) : r(x), g(y), b(z) {}
     __host__ __device__ Rgb(cv::Vec3b v) : r(v[0]), g(v[1]), b(v[2]) {}
     __host__ __device__ void div(int x)
     {
@@ -12,13 +12,13 @@ struct Rgb {
         g /= x;
         b /= x;
     }
-    float r;
-    float g;
-    float b;
+    double r;
+    double g;
+    double b;
 };
 
 __global__ void knn(Rgb* device_img, Rgb* img, int width, int height, int conv_size, double h_param);
-__global__ void non_local_means_gpu(Rgb* device_img, Rgb* img, int conv_size, float weight_decay);
+__global__ void non_local_means_gpu(Rgb* device_img, Rgb* img, int conv_size, double weight_decay);
 __global__ void kernel_shared_conv(Rgb* device_img, Rgb* img, int width, int height, int strel_size);
 __global__ void kernel_conv(Rgb* device_img, Rgb* img, int rows, int cols, int conv_size);
 __global__ void kernel_pixelize(Rgb* device_img, Rgb* img, int rows, int cols, int conv_size);
